requirements_api.txt

fastapi
uvicorn[standard]
redis
requests
joblib
scikit-learn
prometheus-fastapi-instrumentator
python-dotenv
pydantic
psycopg2-binary